Patch provided by Kevin Salvesen:
===============================
Title & subtitle warning Qt bug
===============================

Summary
-------
When underlining a string with '=' or '-' characters an exception is raised which causes the Qt Webkit 
control to stop updating.

Details
-------
Whenever I type exactly 4 times either '=' or '-', the bug happens. 

The exception is not raised when underlining a string of length 1, 2 or 3, even when using 4 or more 
underline characters. 

This is the raised exception::

    Exception in thread Thread-2:
    Traceback (most recent call last):
      File "/usr/lib/python2.6/threading.py", line 525, in __bootstrap_inner
        self.run()
      File "/usr/lib/python2.6/threading.py", line 477, in run
        self.__target(*self.__args, **self.__kwargs)
      File "/usr/lib/python2.6/multiprocessing/pool.py", line 275, in _handle_results
        cache[job]._set(i, obj)
      File "/usr/lib/python2.6/multiprocessing/pool.py", line 427, in _set
        self._callback(self._value)
      File "/home/kevin/ETS_3.3.1/AppTools/enthought/rst/rest_editor_model.py", line 103, in _set_html
        description = node.children[0].children[0].data
    AttributeError: 'Text' object has no attribute 'data'


Where node.children[0].children[0] contains strings like:

- "Document or section may not begin with a transition."
- "Document may not end with a transition."
- "Invalid section title or transition marker."
    
Fix
---
The problem was in line 103 in rest_editor_model.py, where the program was trying to access a 
non existing 'data' attribute::

    description = node.children[0].children[0].data

The fix was to replace that line with::

    description = node.children[0].children[0]
    
which contains the appropriate warning text. With this change, the warnings show up properly 
in the warning tab of the reStructured Text Editor.