TEST: Use .txt files instead of .py files for test
Background: On Windows, Python-2.7.6 has changed the way the mimetypes
are computed and for a .py file `mimetypes.guess_type` now returns
`('text/plain', None)` instead of `('text/x-python', None)`.  See here:
http://bugs.python.org/issue15207 This causes the tests to fail on
Windows alone.  The test now uses a `.txt` file so as to side-step this
issue for the time being.