Fix idempotency bug misrepresenting cache usage
The implementation ported from python 3 had slightly
different semantics than what we need here. Specifically,
inserting twice on a key caused that key to consume two
cache slots in the underlying circular linked list.

The implementation was complicated enough that it wasn't
worth the debugging time to fully dissect the issue. This
implementation is based on collections.OrderedDict and is
considerably simpler to reason about.