Add workflow to publish releases to PyPI (#315)
This PR adds a workflow which automatically publishes releases to PyPI.

This is almost identical to the corresponding workflow on Envisage (which ran successfully this morning). The only difference is an extra line to upgrade pip if necessary. (This line doesn't make a difference right now, but it's not inconceivable that some future release of build or twine can only be installed properly with an updated pip.)